Roughing out some plumbing
The RuleEngineException is the base for all the ways the execution can fail, each carrying an ExitCode enum value that translates to an exit code. This way objects can use exceptions as usual, without worrying about return codes; the exception can be turned into an exit code by the app higher up.